chore: move types features into dedicated module
